moving from mod to assert
